requirements.txt
PyYAML
python-telegram-bot==20.0.0
urllib3==1.26.15
six==1.16.0
requests
